Change gemm to double values

+++ include/gemm.h
@@ -1,6 +1,6 @@
 #ifndef GEMM_H_
 #define GEMM_H_
 
-void gemm(int n, int* c, int* a, int* b); 
+void gemm(int n, double* c, double* a, double* b); 
 
 #endif  // GEMM_H_
